fix(conform-react): support resolveSubmission on server
