+++ .changeset/calm-forms-resolve.md
@@ -1,0 +1,5 @@
+---
+'@conform-to/react': patch
+---
+
+fix: make `resolveSubmission` available in React Server Components
